Feature: Add index all data step (#136)
* move the submission_all_data concern to a submission process service

* add index_all step to submission parsing steps

* add index all data submission status
