+++ lib/ontologies_linked_data/models/submission_status.rb
@@ -7,6 +7,7 @@ module LinkedData
         "RDF_LABELS", "ERROR_RDF_LABELS",
         "OBSOLETE", "ERROR_OBSOLETE",
         "INDEXED", "ERROR_INDEXED",
+        "INDEXED_ALL_DATA", "ERROR_INDEXED_ALL_DATA",
         "INDEXED_PROPERTIES", "ERROR_INDEXED_PROPERTIES",
         "METRICS",  "ERROR_METRICS",
         "ANNOTATOR", "ERROR_ANNOTATOR",
@@ -18,6 +19,8 @@ module LinkedData
         "RDF"                      => "Parsed successfully",
         "RDF_ERROR"                => "Error parsing",
         "INDEXED"                  => "Indexed terms for search",
+        "INDEXED_ALL_DATA"         => "Indexed all the data of the resource",
+        "ERROR_INDEXED_ALL_DATA"   => "Error indexeding all the data of the resource",
         "ERROR_INDEXED"            => "Error indexing terms for search",
         "INDEXED_PROPERTIES"       => "Indexed properties for search",
         "ERROR_INDEXED_PROPERTIES" => "Error indexing properties for search",
